updated working files

+++ ft_strnstr.c
@@ -6,7 +6,7 @@
 /*   By: ahartley <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/05/21 11:06:30 by ahartley          #+#    #+#             */
-/*   Updated: 2019/05/23 10:45:52 by ahartley         ###   ########.fr       */
+/*   Updated: 2019/05/24 15:20:20 by ahartley         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,9 +19,9 @@ char	*ft_strnstr(const char *haystack, const char *needle, size_t len)
 
 	i = 0;
 	j = 0;
-	while ((haystack[i]) && (i < len))
+	while ((haystack[i]) && (i < (int)len))
 	{
-		while ((needle[j] == haystack[i]) && (i < len))
+		while ((needle[j] == haystack[i]) && (i < (int)len))
 		{
 			i++;
 			j++;
